+++ Quicksilver/Code-QuickStepCore/QSCatalogEntry.h
@@ -35,8 +35,8 @@
 - (BOOL)isPreset;
 - (BOOL)isSeparator;
 - (BOOL)isGroup;
-- (int) state;
-- (int) hasEnabledChildren;
+- (NSInteger) state;
+- (NSInteger) hasEnabledChildren;
 - (BOOL)isEnabled;
 - (void)setEnabled:(BOOL)enabled;
 - (void)setDeepEnabled:(BOOL)enabled;
@@ -48,7 +48,7 @@
 - (NSArray *)ancestors;
 - (NSString *)name;
 - (NSImage *)icon;
-- (int) deepObjectCount;
+- (NSInteger) deepObjectCount;
 - (BOOL)loadIndex;
 - (void)saveIndex;
 - (BOOL)indexIsValid;
@@ -77,6 +77,6 @@
 - (void)setIsScanning:(BOOL)flag;
 //- (NSString *)countString;
 
-- (int) count;
+- (NSInteger) count;
 - (NSIndexPath *)catalogSetIndexPath;
 @end
