+++ Quicksilver/Code-QuickStepCore/QSCatalogEntry.m
@@ -156,7 +156,7 @@ NSDictionary *enabledPresetDictionary;*/
 - (BOOL)isSeparator { return [[self identifier] hasPrefix:@"QSSeparator"]; }
 - (BOOL)isGroup { return [[info objectForKey:kItemSource] isEqualToString:@"QSGroupObjectSource"]; }
 - (BOOL)isLeaf { return ![self isGroup]; }
-- (int)state {
+- (NSInteger)state {
 	BOOL enabled = [self isEnabled];
 	if (!enabled) return 0;
 	if ([[info objectForKey:kItemSource] isEqualToString:@"QSGroupObjectSource"]) {
@@ -167,7 +167,7 @@ NSDictionary *enabledPresetDictionary;*/
 	return enabled;
 }
 
-- (int)hasEnabledChildren {
+- (NSInteger)hasEnabledChildren {
 	if ([[info objectForKey:kItemSource] isEqualToString:@"QSGroupObjectSource"]) {
 		BOOL hasEnabledChildren = NO;
 		for (id loopItem in children)
@@ -283,8 +283,8 @@ NSDictionary *enabledPresetDictionary;*/
 
 - (NSIndexPath *)catalogIndexPath {
 	NSArray *anc = [self ancestors];
-	int i;
-	int index;
+	NSInteger i;
+	NSInteger index;
 	NSIndexPath *p = nil;
 	for (i = 0; i < ([anc count]-1); i++) {
 		index = [[[anc objectAtIndex:i] children] indexOfObject:[anc objectAtIndex:i+1]];
@@ -299,8 +299,8 @@ NSDictionary *enabledPresetDictionary;*/
 
 - (NSIndexPath *)catalogSetIndexPath {
 	NSArray *anc = [self ancestors];
-	int i;
-	int index;
+	NSInteger i;
+	NSInteger index;
 	NSIndexPath *p = nil;
 	for (i = 1; i<([anc count] -1); i++) {
 		index = [[[anc objectAtIndex:i] children] indexOfObject:[anc objectAtIndex:i+1]];
@@ -318,7 +318,7 @@ NSDictionary *enabledPresetDictionary;*/
 	NSArray *groups = [catalog deepChildrenWithGroups:YES leaves:NO disabled:YES];
 	NSMutableArray *entryChain = [NSMutableArray arrayWithCapacity:0];
 	id thisItem = self;
-	int i;
+	NSInteger i;
 	[entryChain addObject:self];
 	id theGroup = nil;
 	while(thisItem != catalog) {
@@ -378,20 +378,20 @@ NSDictionary *enabledPresetDictionary;*/
 }
 
 - (NSString *)getCount {
-	int num;
+	NSInteger num;
 	if((num = [self count]))
 		return [NSString stringWithFormat:@"%d", num];
 	else
 		return nil;
 }
 
-- (int)count {
+- (NSInteger)count {
 	return [self deepObjectCount];
 }
 
-- (int)deepObjectCount {
+- (NSInteger)deepObjectCount {
 	NSArray *leaves = [self deepChildrenWithGroups:NO leaves:YES disabled:NO];
-	int i, count = 0;
+	NSInteger i, count = 0;
 	for (i = 0; i<[leaves count]; i++)
 		count += [(NSArray *)[[leaves objectAtIndex:i] contents] count];
 	return count;
