+++ Quicksilver/Code-QuickStepFoundation/QSLog.m
@@ -9,14 +9,14 @@
 #import "QSLog.h"
 
 @implementation QSLog
-+(void)logFile:(char*)sourceFile lineNumber:(int)lineNumber format:(NSString*)format, ... {
++(void)logFile:(char*)sourceFile lineNumber:(NSInteger)lineNumber format:(NSString*)format, ... {
 	va_list ap;
 	NSString *print, *file;
 	va_start(ap, format);
 	file = [[NSString alloc] initWithBytes:sourceFile length:strlen(sourceFile) encoding:NSUTF8StringEncoding];
 	print = [[NSString alloc] initWithFormat:format arguments:ap];
 	va_end(ap);
-	NSLog(@"%s:%d %@", [[file lastPathComponent] UTF8String], lineNumber, print);
+	NSLog(@"%s:%ld %@", [[file lastPathComponent] UTF8String], lineNumber, print);
 	[print release];
 	[file release];
 }
